+++ tests/Tests/Parametizer/EnvironmentConfig/autoload/level2/level3/no-search-above-topmost.php
@@ -3,15 +3,9 @@
 declare(strict_types=1);
 
 use MagicPush\CliToolkit\Parametizer\EnvironmentConfig;
+use MagicPush\CliToolkit\Tests\utils\TestUtils;
 
 require_once __DIR__ . '/../../../../../init-console.php';
 
 $envConfig = EnvironmentConfig::createFromConfigsBottomUpHierarchy(__DIR__, __DIR__, true);
-echo json_encode(
-    $envConfig,
-    JSON_THROW_ON_ERROR
-    | JSON_UNESCAPED_UNICODE
-    | JSON_UNESCAPED_SLASHES
-    | JSON_UNESCAPED_LINE_TERMINATORS
-    | JSON_PRETTY_PRINT,
-);
+echo TestUtils::getEnvironmentConfigPartJson($envConfig);