+++ docs/changelog.md
@@ -9,22 +9,63 @@ This change log references the repository changes and releases, which respect [s
 1. [cli-toolkit](../tools/cli-toolkit) plain scripts are removed to be replaced with
    [ScriptAbstract.php](../src/Parametizer/Script/ScriptAbstract.php)-based scripts
    and a [launcher.php](../tools/cli-toolkit/launcher.php).
+1. `HelpGenerator::getSubcommandsBlock()` is removed - replaced with `list` built-in subcommand.
+1. `CliRequest::getSubcommandRequest()`: removed `$subcommandName` parameter, changed return type hint
+   from `string` to `?string`:
+   now the method always returns the called subcommand's request object or `null` if there was no subcommand call.
+1. [HelpGenerator.php](../src/Parametizer/Config/HelpGenerator.php):
+    1. Removed `getSubcommandsBlock()` (so as "COMMANDS" block output from `getFullHelp()`)
+       to replace it with `list` built-in subcommand functionality.
+    1. Removed `getBaseScriptName()` obsolete method.
 
 ### New features
 
+1. Removed "minimum 2 subcommands" constraint from `Config::commitSubcommandSwitch()`.
 1. [ScriptAbstract.php](../src/Parametizer/Script/ScriptAbstract.php) as a basement for
    class-based Parametizer-powered scripts.
-1. [CliToolkitScriptAbstract.php](../tools/cli-toolkit/Scripts/CliToolkitScriptAbstract.php) as a basement for all
-   [tools/cli-toolkit](../tools/cli-toolkit) scripts.
-1. [ScriptDetector.php](../src/Parametizer/Script/ScriptDetector.php) for different script types auto-detection.
-   For now only [ScriptAbstract.php](../src/Parametizer/Script/ScriptAbstract.php)-based scripts are supported.
-1. [GenerateMassTestScripts.php](../tools/cli-toolkit/Scripts/Internal/GenerateMassTestScripts.php) as a tool
-   to test the performance and other "law of large numbers" cases, when a launcher includes lots of class scripts.
-1. `ScriptFormatter::note()` is added initially for
-   [GenerateMassTestScripts.php](../tools/cli-toolkit/Scripts/Internal/GenerateMassTestScripts.php)
 1. Subcommand names (`Config::newSubcommand()`) now support the colon (`:`) symbol.
    Main purpose - a separator for script classes sections.
-1. Removed min 2 subcommands constraint from `Config::commitSubcommandSwitch()`.
+1. [ScriptDetector.php](../src/Parametizer/Script/ScriptDetector.php) for different script types auto-detection.
+   For now only [ScriptAbstract.php](../src/Parametizer/Script/ScriptAbstract.php)-based scripts are supported.
+1. Built-in subcommands: each script with a subcommand switch automatically provides you with
+   `help` ([HelpScript.php](../src/Parametizer/Script/BuiltInSubcommand/HelpScript.php))
+   and `list` ([ListScript.php](../src/Parametizer/Script/BuiltInSubcommand/ListScript.php)) built-in subcommands.
+1. [VariableBuilderAbstract.php](../src/Parametizer/Config/Builder/VariableBuilderAbstract.php):
+    1. Added an optional `$areHiddenForHelp` parameter for `allowedValues()` method, defaults to `false`.
+    1. Added a protected method `setAllowedValues()` for common internal operations with the allowed values list.
+1. [ParameterAbstract.php](../src/Parametizer/Config/Parameter/ParameterAbstract.php):
+   added a protected property `$areAllowedValuesHiddenFromHelp`,
+   a related getter method `areAllowedValuesHiddenFromHelp()`,
+   and a related optional parameter `$areHiddenFromHelp` for `allowedValues()` method.
+1. `HelpGenerator::makeParamDescription()`:
+    * Utilizes `$areAllowedValuesHiddenFromHelp` parameter property and does not show the list of values
+      if the flag is set to `true`.
+    * For subcommands: replaces "Allowed values" actual list with a hint about detected subcommands count
+      and a subcommand name to show all available subcommands.
+1. [CliRequest.php](../src/Parametizer/CliRequest/CliRequest.php):
+    1. `__construct()` changes:
+        1. `$config` parameter is made _public_ (from _protected_).
+        1. Added `$parent` parameter to access parent request (from a subcommand request).
+    1. Added `getSubcommandRequestName()` method.
+    1. Added `executeBuiltInSubcommandIfRequested()` method for built-in subcommands automatic execution;
+       the method is utilized by `Parametizer::run()`.
+1. [Config.php](../src/Parametizer/Config/Config.php):
+    1. Added `PARAMETER_NAME_LIST` public constant to keep the listing built-in subcommand name.
+    1. Added `$builtInSubcommandClassBySubcommandName` protected property and the related methods:
+        * public `getSubcommandSwitchName()` - to get a subcommand switch parameter name if present in a config;
+        * public `getBuiltInSubcommandClass()` - to get a subcommand fully qualified class name by a subcommand name;
+        * public `getBuiltInSubcommands()` - to get the list of built-in subcommand configs by their names;
+        * protected `addBuiltInSubcommands()` - an internal method to add built-in subcommands automatically to every
+          script that contains a subcommand switch parameter.
+1. [CliToolkitScriptAbstract.php](../tools/cli-toolkit/Scripts/CliToolkitScriptAbstract.php) as a basement for all
+   [tools/cli-toolkit](../tools/cli-toolkit) scripts.
+1. [GenerateMassTestScripts.php](../tools/cli-toolkit/Scripts/Internal/GenerateMassTestScripts.php) as a tool
+   to test the performance and other "law of large numbers" cases, when a launcher includes lots of class scripts.
+1. Formatters:
+    1. Added `HelpFormatter::invert()`.
+    1. Added `ScriptFormatter::note()`.
+1. Added `$dieMessage` parameter to `Question::confirmOrDie()` method - outputs a message (if provided)
+   before interrupting script's execution.
 
 ## v2.1.0
 
@@ -46,7 +87,7 @@ This change log references the repository changes and releases, which respect [s
    in your production classes under 'dev' environment (when you call `composer install` without `--no-dev` option),
    you will get "Class 'XXX' not found in ..." error.
    Previously there was no error, until you install composer packages with `--no-dev` flag.
-1. `HelpGenerator`:
+1. [HelpGenerator.php](../src/Parametizer/Config/HelpGenerator.php):
     1. Fixed scripts main description block - stopped counting symbols in space-only lines.
        Previously it caused too much padding for descriptions with too short space-only lines.
     1. Improved subcommand help usage block - when `--help` is called for a subcommand, all manual usage lines
@@ -69,7 +110,7 @@ This change log references the repository changes and releases, which respect [s
 1. `Parametizer::setExceptionHandlerForParsing()` renders both an error message and a help block in STDERR (previously
    a help block was rendered in STDOUT).
 1. `Option::getOptionCleanFullName()` is deleted because of no usage.
-   
+
 1. Renaming:
     1. `CliRequest::getCommandRequest()` -> `getSubcommandRequest()`
     1. `CliRequestProcessor::getAllowedArguments()` -> `getInnermostBranchAllowedArguments()`